+++ src/Language/Futhark/TypeChecker/Terms.hs
@@ -25,6 +25,7 @@ import Control.Monad.RWS hiding (Sum)
 import Control.Monad.State
 import Control.Monad.Writer hiding (Sum)
 import Data.Bifunctor
+import Data.Bitraversable
 import Data.Char (isAscii)
 import Data.Either
 import Data.List (find, foldl', isPrefixOf, sort)
@@ -664,7 +665,7 @@ instantiateTypeParam loc tparam = do
   case tparam of
     TypeParamType x _ _ -> do
       constrain v $ NoConstraint x $ mkUsage' loc
-      return (v, Subst $ Scalar $ TypeVar mempty Nonunique (typeName v) [])
+      return (v, Subst [] $ Scalar $ TypeVar mempty Nonunique (typeName v) [])
     TypeParamDim {} -> do
       constrain v $ Size Nothing $ mkUsage' loc
       return (v, SizeSubst $ NamedDim $ qualName v)
@@ -768,56 +769,73 @@ patLitMkType (PatLitFloat _) loc = do
 patLitMkType (PatLitPrim v) _ =
   pure $ Scalar $ Prim $ primValueType v
 
+nonrigidFor :: [SizeBinder VName] -> StructType -> TermTypeM StructType
+nonrigidFor [] t = pure t -- Minor optimisation.
+nonrigidFor sizes t = evalStateT (bitraverse onDim pure t) mempty
+  where
+    onDim (NamedDim (QualName _ v))
+      | Just size <- find ((== v) . sizeName) sizes = do
+        prev <- gets $ lookup v
+        case prev of
+          Nothing -> do
+            v' <- lift $ newID $ baseName v
+            lift $ constrain v' $ Size Nothing $ mkUsage' $ srclocOf size
+            modify ((v, v') :)
+            pure $ NamedDim $ qualName v'
+          Just v' ->
+            pure $ NamedDim $ qualName v'
+    onDim d = pure d
+
 checkPattern' ::
+  [SizeBinder VName] ->
   UncheckedPattern ->
   InferredType ->
   TermTypeM Pattern
-checkPattern' (PatternParens p loc) t =
-  PatternParens <$> checkPattern' p t <*> pure loc
-checkPattern' (Id name _ loc) _
+checkPattern' sizes (PatternParens p loc) t =
+  PatternParens <$> checkPattern' sizes p t <*> pure loc
+checkPattern' _ (Id name _ loc) _
   | name' `elem` doNotShadow =
     typeError loc mempty $ "The" <+> text name' <+> "operator may not be redefined."
   where
     name' = nameToString name
-checkPattern' (Id name NoInfo loc) (Ascribed t) = do
+checkPattern' _ (Id name NoInfo loc) (Ascribed t) = do
   name' <- newID name
   return $ Id name' (Info t) loc
-checkPattern' (Id name NoInfo loc) NoneInferred = do
+checkPattern' _ (Id name NoInfo loc) NoneInferred = do
   name' <- newID name
   t <- newTypeVar loc "t"
   return $ Id name' (Info t) loc
-checkPattern' (Wildcard _ loc) (Ascribed t) =
+checkPattern' _ (Wildcard _ loc) (Ascribed t) =
   return $ Wildcard (Info $ t `setUniqueness` Nonunique) loc
-checkPattern' (Wildcard NoInfo loc) NoneInferred = do
+checkPattern' _ (Wildcard NoInfo loc) NoneInferred = do
   t <- newTypeVar loc "t"
   return $ Wildcard (Info t) loc
-checkPattern' (TuplePattern ps loc) (Ascribed t)
+checkPattern' sizes (TuplePattern ps loc) (Ascribed t)
   | Just ts <- isTupleRecord t,
     length ts == length ps =
-    TuplePattern <$> zipWithM checkPattern' ps (map Ascribed ts) <*> pure loc
-checkPattern' p@(TuplePattern ps loc) (Ascribed t) = do
+    TuplePattern
+      <$> zipWithM (checkPattern' sizes) ps (map Ascribed ts)
+      <*> pure loc
+checkPattern' sizes p@(TuplePattern ps loc) (Ascribed t) = do
   ps_t <- replicateM (length ps) (newTypeVar loc "t")
   unify (mkUsage loc "matching a tuple pattern") (tupleRecord ps_t) $ toStruct t
   t' <- normTypeFully t
-  checkPattern' p $ Ascribed t'
-checkPattern' (TuplePattern ps loc) NoneInferred =
-  TuplePattern <$> mapM (`checkPattern'` NoneInferred) ps <*> pure loc
-checkPattern' (RecordPattern p_fs _) _
+  checkPattern' sizes p $ Ascribed t'
+checkPattern' sizes (TuplePattern ps loc) NoneInferred =
+  TuplePattern <$> mapM (\p -> checkPattern' sizes p NoneInferred) ps <*> pure loc
+checkPattern' _ (RecordPattern p_fs _) _
   | Just (f, fp) <- find (("_" `isPrefixOf`) . nameToString . fst) p_fs =
     typeError fp mempty $
       "Underscore-prefixed fields are not allowed."
         </> "Did you mean" <> dquotes (text (drop 1 (nameToString f)) <> "=_") <> "?"
-checkPattern' (RecordPattern p_fs loc) (Ascribed (Scalar (Record t_fs)))
+checkPattern' sizes (RecordPattern p_fs loc) (Ascribed (Scalar (Record t_fs)))
   | sort (map fst p_fs) == sort (M.keys t_fs) =
     RecordPattern . M.toList <$> check <*> pure loc
   where
     check =
-      traverse (uncurry checkPattern') $
-        M.intersectionWith
-          (,)
-          (M.fromList p_fs)
-          (fmap Ascribed t_fs)
-checkPattern' p@(RecordPattern fields loc) (Ascribed t) = do
+      traverse (uncurry (checkPattern' sizes)) $
+        M.intersectionWith (,) (M.fromList p_fs) (fmap Ascribed t_fs)
+checkPattern' sizes p@(RecordPattern fields loc) (Ascribed t) = do
   fields' <- traverse (const $ newTypeVar loc "t") $ M.fromList fields
 
   when (sort (M.keys fields') /= sort (map fst fields)) $
@@ -825,17 +843,20 @@ checkPattern' p@(RecordPattern fields loc) (Ascribed t) = do
 
   unify (mkUsage loc "matching a record pattern") (Scalar (Record fields')) $ toStruct t
   t' <- normTypeFully t
-  checkPattern' p $ Ascribed t'
-checkPattern' (RecordPattern fs loc) NoneInferred =
-  RecordPattern . M.toList <$> traverse (`checkPattern'` NoneInferred) (M.fromList fs) <*> pure loc
-checkPattern' (PatternAscription p (TypeDecl t NoInfo) loc) maybe_outer_t = do
+  checkPattern' sizes p $ Ascribed t'
+checkPattern' sizes (RecordPattern fs loc) NoneInferred =
+  RecordPattern . M.toList
+    <$> traverse (\p -> checkPattern' sizes p NoneInferred) (M.fromList fs)
+    <*> pure loc
+checkPattern' sizes (PatternAscription p (TypeDecl t NoInfo) loc) maybe_outer_t = do
   (t', st_nodims, _) <- checkTypeExp t
   (st, _) <- instantiateEmptyArrayDims loc "impl" Nonrigid st_nodims
 
   let st' = fromStruct st
   case maybe_outer_t of
     Ascribed outer_t -> do
-      unify (mkUsage loc "explicit type ascription") (toStruct st) (toStruct outer_t)
+      st_forunify <- nonrigidFor sizes st
+      unify (mkUsage loc "explicit type ascription") st_forunify (toStruct outer_t)
 
       -- We also have to make sure that uniqueness matches.  This is
       -- done explicitly, because it is ignored by unification.
@@ -843,7 +864,7 @@ checkPattern' (PatternAscription p (TypeDecl t NoInfo) loc) maybe_outer_t = do
       outer_t' <- normTypeFully outer_t
       case unifyTypesU unifyUniqueness st'' outer_t' of
         Just outer_t'' ->
-          PatternAscription <$> checkPattern' p (Ascribed outer_t'')
+          PatternAscription <$> checkPattern' sizes p (Ascribed outer_t'')
             <*> pure (TypeDecl t' (Info st))
             <*> pure loc
         Nothing ->
@@ -851,33 +872,33 @@ checkPattern' (PatternAscription p (TypeDecl t NoInfo) loc) maybe_outer_t = do
             "Cannot match type" <+> pquote (ppr outer_t') <+> "with expected type"
               <+> pquote (ppr st'') <> "."
     NoneInferred ->
-      PatternAscription <$> checkPattern' p (Ascribed st')
+      PatternAscription <$> checkPattern' sizes p (Ascribed st')
         <*> pure (TypeDecl t' (Info st))
         <*> pure loc
   where
     unifyUniqueness u1 u2 = if u2 `subuniqueOf` u1 then Just u1 else Nothing
-checkPattern' (PatternLit l NoInfo loc) (Ascribed t) = do
+checkPattern' _ (PatternLit l NoInfo loc) (Ascribed t) = do
   t' <- patLitMkType l loc
   unify (mkUsage loc "matching against literal") t' (toStruct t)
   return $ PatternLit l (Info (fromStruct t')) loc
-checkPattern' (PatternLit l NoInfo loc) NoneInferred = do
+checkPattern' _ (PatternLit l NoInfo loc) NoneInferred = do
   t' <- patLitMkType l loc
   return $ PatternLit l (Info (fromStruct t')) loc
-checkPattern' (PatternConstr n NoInfo ps loc) (Ascribed (Scalar (Sum cs)))
+checkPattern' sizes (PatternConstr n NoInfo ps loc) (Ascribed (Scalar (Sum cs)))
   | Just ts <- M.lookup n cs = do
-    ps' <- zipWithM checkPattern' ps $ map Ascribed ts
+    ps' <- zipWithM (checkPattern' sizes) ps $ map Ascribed ts
     return $ PatternConstr n (Info (Scalar (Sum cs))) ps' loc
-checkPattern' (PatternConstr n NoInfo ps loc) (Ascribed t) = do
+checkPattern' sizes (PatternConstr n NoInfo ps loc) (Ascribed t) = do
   t' <- newTypeVar loc "t"
-  ps' <- mapM (`checkPattern'` NoneInferred) ps
+  ps' <- mapM (\p -> checkPattern' sizes p NoneInferred) ps
   mustHaveConstr usage n t' (patternStructType <$> ps')
   unify usage t' (toStruct t)
   t'' <- normTypeFully t
   return $ PatternConstr n (Info t'') ps' loc
   where
     usage = mkUsage loc "matching against constructor"
-checkPattern' (PatternConstr n NoInfo ps loc) NoneInferred = do
-  ps' <- mapM (`checkPattern'` NoneInferred) ps
+checkPattern' sizes (PatternConstr n NoInfo ps loc) NoneInferred = do
+  ps' <- mapM (\p -> checkPattern' sizes p NoneInferred) ps
   t <- newTypeVar loc "t"
   mustHaveConstr usage n t (patternStructType <$> ps')
   return $ PatternConstr n (Info $ fromStruct t) ps' loc
@@ -890,14 +911,24 @@ patternNameMap = M.fromList . map asTerm . S.toList . patternNames
     asTerm v = ((Term, baseName v), qualName v)
 
 checkPattern ::
+  [SizeBinder VName] ->
   UncheckedPattern ->
   InferredType ->
   (Pattern -> TermTypeM a) ->
   TermTypeM a
-checkPattern p t m = do
+checkPattern sizes p t m = do
   checkForDuplicateNames [p]
-  p' <- onFailure (CheckingPattern p t) $ checkPattern' p t
-  bindNameMap (patternNameMap p') $ m p'
+  p' <- onFailure (CheckingPattern p t) $ checkPattern' sizes p t
+
+  let explicit = mustBeExplicitInType $ patternStructType p'
+
+  case filter ((`S.member` explicit) . sizeName) sizes of
+    size : _ ->
+      typeError size mempty $
+        "Cannot bind" <+> ppr size
+          <+> "as it is never used as the size of a concrete (non-function) value."
+    [] ->
+      bindNameMap (patternNameMap p') $ m p'
 
 binding :: [Ident] -> TermTypeM a -> TermTypeM a
 binding bnds = check . handleVars
@@ -1022,7 +1053,7 @@ bindingParams tps orig_ps m = do
   checkForDuplicateNames orig_ps
   checkTypeParams tps $ \tps' -> bindingTypeParams tps' $ do
     let descend ps' (p : ps) =
-          checkPattern p NoneInferred $ \p' ->
+          checkPattern [] p NoneInferred $ \p' ->
             binding (S.toList $ patternIdents p') $ descend (p' : ps') ps
         descend ps' [] = do
           -- Perform an observation of every type parameter.  This
@@ -1033,19 +1064,49 @@ bindingParams tps orig_ps m = do
 
     descend [] orig_ps
 
+bindingSizes :: [SizeBinder Name] -> ([SizeBinder VName] -> TermTypeM a) -> TermTypeM a
+bindingSizes [] m = m [] -- Minor optimisation.
+bindingSizes sizes m = do
+  foldM_ lookForDuplicates mempty sizes
+  bindSpaced (map sizeWithSpace sizes) $ do
+    sizes' <- mapM check sizes
+    binding (map sizeWithType sizes') $ m sizes'
+  where
+    lookForDuplicates prev size
+      | Just prevloc <- M.lookup (sizeName size) prev =
+        typeError size mempty $
+          "Size name also bound at "
+            <> text (locStrRel (srclocOf size) prevloc)
+            <> "."
+      | otherwise =
+        pure $ M.insert (sizeName size) (srclocOf size) prev
+
+    sizeWithSpace size =
+      (Term, sizeName size)
+    sizeWithType size =
+      Ident (sizeName size) (Info (Scalar (Prim (Signed Int64)))) (srclocOf size)
+
+    check (SizeBinder v loc) =
+      SizeBinder <$> checkName Term v loc <*> pure loc
+
 bindingPattern ::
+  [SizeBinder VName] ->
   PatternBase NoInfo Name ->
   InferredType ->
   (Pattern -> TermTypeM a) ->
   TermTypeM a
-bindingPattern p t m = do
+bindingPattern sizes p t m = do
   checkForDuplicateNames [p]
-  checkPattern p t $ \p' -> binding (S.toList $ patternIdents p') $ do
+  checkPattern sizes p t $ \p' -> binding (S.toList $ patternIdents p') $ do
     -- Perform an observation of every declared dimension.  This
     -- prevents unused-name warnings for otherwise unused dimensions.
     mapM_ observe $ patternDims p'
 
-    m p'
+    let used_sizes = typeDimNames $ patternStructType p'
+    case filter ((`S.notMember` used_sizes) . sizeName) sizes of
+      [] -> m p'
+      size : _ ->
+        typeError size mempty $ "Size" <+> ppr size <+> "unused in pattern."
 
 patternDims :: Pattern -> [Ident]
 patternDims (PatternParens p _) = patternDims p
@@ -1175,7 +1236,7 @@ checkAscript loc decl e shapef = do
   -- explicitly, because uniqueness is ignored by unification.
   t' <- normTypeFully t
   decl_t' <- normTypeFully $ unInfo $ expandedType decl'
-  unless (noSizes t' `subtypeOf` noSizes decl_t') $
+  unless (toStructural t' `subtypeOf` toStructural decl_t') $
     typeError loc mempty $
       "Type" <+> pquote (ppr t') <+> "is not a subtype of"
         <+> pquote (ppr decl_t') <> "."
@@ -1451,7 +1512,7 @@ checkExp (Negate arg loc) = do
   arg' <- require "numeric negation" anyNumberType =<< checkExp arg
   return $ Negate arg' loc
 checkExp e@(AppExp Apply {} _) = fst <$> checkApplyExp e
-checkExp (AppExp (LetPat pat e body loc) _) =
+checkExp (AppExp (LetPat sizes pat e body loc) _) =
   sequentially (checkExp e) $ \e' e_occs -> do
     -- Not technically an ascription, but we want the pattern to have
     -- exactly the type of 'e'.
@@ -1462,13 +1523,18 @@ checkExp (AppExp (LetPat pat e body loc) _) =
          in zeroOrderType (mkUsage loc "consumption in right-hand side of 'let'-binding") msg t
       _ -> return ()
 
-    incLevel $
-      bindingPattern pat (Ascribed t) $ \pat' -> do
+    incLevel . bindingSizes sizes $ \sizes' ->
+      bindingPattern sizes' pat (Ascribed t) $ \pat' -> do
         body' <- checkExp body
         (body_t, retext) <-
-          unscopeType loc (patternMap pat') =<< expTypeFully body'
+          unscopeType loc (sizesMap sizes' <> patternMap pat') =<< expTypeFully body'
 
-        return $ AppExp (LetPat pat' e' body' loc) (Info $ AppRes body_t retext)
+        return $ AppExp (LetPat sizes' pat' e' body' loc) (Info $ AppRes body_t retext)
+  where
+    sizesMap = foldMap onSize
+    onSize size =
+      M.singleton (sizeName size) $
+        Ident (sizeName size) (Info (Scalar $ Prim $ Signed Int64)) (srclocOf size)
 checkExp (AppExp (LetFun name (tparams, params, maybe_retdecl, NoInfo, e) body loc) _) =
   sequentially (checkBinding (name, maybe_retdecl, tparams, params, e, loc)) $
     \(tparams', params', maybe_retdecl', rettype, _, e') closure -> do
@@ -1762,10 +1828,11 @@ checkExp (AppExp (DoLoop _ mergepat mergeexp form loopbody loc) _) =
           pat_t <- normTypeFully $ patternType mergepat'
           -- We are ignoring the dimensions here, because any mismatches
           -- should be turned into fresh size variables.
+
           onFailure (CheckingLoopBody (toStruct (anySizes pat_t)) (toStruct loopbody_t)) $
             expect
               (mkUsage (srclocOf loopbody) "matching loop body to loop pattern")
-              (toStruct (anySizes pat_t))
+              (toStruct (anyTheseSizes new_dims pat_t))
               (toStruct loopbody_t)
           pat_t' <- normTypeFully pat_t
           loopbody_t' <- normTypeFully loopbody_t
@@ -1832,18 +1899,17 @@ checkExp (AppExp (DoLoop _ mergepat mergeexp form loopbody loc) _) =
           uboundexp' <- require "being the bound in a 'for' loop" anySignedType =<< checkExp uboundexp
           bound_t <- expTypeFully uboundexp'
           bindingIdent i bound_t $ \i' ->
-            noUnique $
-              bindingPattern mergepat (Ascribed merge_t) $
-                \mergepat' -> onlySelfAliasing $
-                  tapOccurences $ do
-                    loopbody' <- noSizeEscape $ checkExp loopbody
-                    (sparams, mergepat'') <- checkLoopReturnSize mergepat' loopbody'
-                    return
-                      ( sparams,
-                        mergepat'',
-                        For i' uboundexp',
-                        loopbody'
-                      )
+            noUnique . bindingPattern [] mergepat (Ascribed merge_t) $
+              \mergepat' -> onlySelfAliasing $
+                tapOccurences $ do
+                  loopbody' <- noSizeEscape $ checkExp loopbody
+                  (sparams, mergepat'') <- checkLoopReturnSize mergepat' loopbody'
+                  return
+                    ( sparams,
+                      mergepat'',
+                      For i' uboundexp',
+                      loopbody'
+                    )
         ForIn xpat e -> do
           (arr_t, _) <- newArrayType (srclocOf e) "e" 1
           e' <- unifies "being iterated in a 'for-in' loop" arr_t =<< checkExp e
@@ -1851,41 +1917,37 @@ checkExp (AppExp (DoLoop _ mergepat mergeexp form loopbody loc) _) =
           case t of
             _
               | Just t' <- peelArray 1 t ->
-                bindingPattern xpat (Ascribed t') $ \xpat' ->
-                  noUnique $
-                    bindingPattern mergepat (Ascribed merge_t) $
-                      \mergepat' -> onlySelfAliasing $
-                        tapOccurences $ do
-                          loopbody' <- noSizeEscape $ checkExp loopbody
-                          (sparams, mergepat'') <- checkLoopReturnSize mergepat' loopbody'
-                          return
-                            ( sparams,
-                              mergepat'',
-                              ForIn xpat' e',
-                              loopbody'
-                            )
-              | otherwise ->
-                typeError (srclocOf e) mempty $
-                  "Iteratee of a for-in loop must be an array, but expression has type"
-                    <+> ppr t
-        While cond ->
-          noUnique $
-            bindingPattern mergepat (Ascribed merge_t) $ \mergepat' ->
-              onlySelfAliasing $
-                tapOccurences $
-                  sequentially
-                    ( checkExp cond
-                        >>= unifies "being the condition of a 'while' loop" (Scalar $ Prim Bool)
-                    )
-                    $ \cond' _ -> do
+                bindingPattern [] xpat (Ascribed t') $ \xpat' ->
+                  noUnique . bindingPattern [] mergepat (Ascribed merge_t) $
+                    \mergepat' -> onlySelfAliasing . tapOccurences $ do
                       loopbody' <- noSizeEscape $ checkExp loopbody
                       (sparams, mergepat'') <- checkLoopReturnSize mergepat' loopbody'
                       return
                         ( sparams,
                           mergepat'',
-                          While cond',
+                          ForIn xpat' e',
                           loopbody'
                         )
+              | otherwise ->
+                typeError (srclocOf e) mempty $
+                  "Iteratee of a for-in loop must be an array, but expression has type"
+                    <+> ppr t
+        While cond ->
+          noUnique . bindingPattern [] mergepat (Ascribed merge_t) $ \mergepat' ->
+            onlySelfAliasing . tapOccurences $
+              sequentially
+                ( checkExp cond
+                    >>= unifies "being the condition of a 'while' loop" (Scalar $ Prim Bool)
+                )
+                $ \cond' _ -> do
+                  loopbody' <- noSizeEscape $ checkExp loopbody
+                  (sparams, mergepat'') <- checkLoopReturnSize mergepat' loopbody'
+                  return
+                    ( sparams,
+                      mergepat'',
+                      While cond',
+                      loopbody'
+                    )
 
     mergepat'' <- do
       loopbody_t <- expTypeFully loopbody'
@@ -1943,6 +2005,12 @@ checkExp (AppExp (DoLoop _ mergepat mergeexp form loopbody loc) _) =
         (DoLoop sparams mergepat'' mergeexp' form' loopbody' loc)
         (Info $ AppRes loopt' retext)
   where
+    anyTheseSizes to_hide = first onDim
+      where
+        onDim (NamedDim (QualName _ v))
+          | v `elem` to_hide = AnyDim Nothing
+        onDim d = d
+
     convergePattern pat body_cons body_t body_loc = do
       let consumed_merge = patternNames pat `S.intersection` body_cons
 
@@ -2090,7 +2158,7 @@ checkCase ::
   CaseBase NoInfo Name ->
   TermTypeM (CaseBase Info VName, PatternType, [VName])
 checkCase mt (CasePat p e loc) =
-  bindingPattern p (Ascribed mt) $ \p' -> do
+  bindingPattern [] p (Ascribed mt) $ \p' -> do
     e' <- checkExp e
     (t, retext) <- unscopeType loc (patternMap p') =<< expTypeFully e'
     return (CasePat p' e' loc, t, retext)
@@ -2265,8 +2333,10 @@ checkApply
       sizeSubst _ _ = return (AnyDim Nothing, Nothing)
 checkApply loc fname tfun@(Scalar TypeVar {}) arg = do
   tv <- newTypeVar loc "b"
+  -- Change the uniqueness of the argument type because we never want
+  -- to infer that a function is consuming.
   unify (mkUsage loc "use as function") (toStruct tfun) $
-    Scalar $ Arrow mempty Unnamed (toStruct (argType arg)) tv
+    Scalar $ Arrow mempty Unnamed (toStruct (argType arg) `setUniqueness` Nonunique) tv
   tfun' <- normPatternType tfun
   checkApply loc fname tfun' arg
 checkApply loc (fname, prev_applied) ftype (argexp, _, _, _) = do
@@ -2440,7 +2510,7 @@ causalityCheck binding_body = do
       onExp known (Lambda params _ _ _ _)
         | bad : _ <- mapMaybe (checkParamCausality known) params =
           bad
-      onExp known e@(AppExp (LetPat _ bindee_e body_e _) (Info res)) = do
+      onExp known e@(AppExp (LetPat _ _ bindee_e body_e _) (Info res)) = do
         sequencePoint known bindee_e body_e $ appResExt res
         return e
       onExp known e@(AppExp (Apply f arg (Info (_, p)) _) (Info res)) = do
@@ -2635,7 +2705,7 @@ fixOverloadedTypes tyvars_at_toplevel =
           <+> ppr (Sum cs) <> ")."
           </> "Add a type annotation to disambiguate the type."
     fixOverloaded (v, Size Nothing usage) =
-      typeError usage mempty $ "Size is ambiguous.\n" <> pprName v
+      typeError usage mempty $ "Size" <+> pquote (pprName v) <+> "is ambiguous.\n"
     fixOverloaded _ = return ()
 
 hiddenParamNames :: [Pattern] -> Names
@@ -3041,7 +3111,7 @@ checkFunBody params body maybe_rettype loc = do
       -- explicitly, because uniqueness is ignored by unification.
       rettype' <- normTypeFully rettype
       body_t'' <- normTypeFully rettype -- Substs may have changed.
-      unless (body_t'' `subtypeOf` anySizes rettype') $
+      unless (toStructural body_t'' `subtypeOf` toStructural rettype') $
         typeError (srclocOf body) mempty $
           "Body type" </> indent 2 (ppr body_t'')
             </> "is not a subtype of annotated type"